feat: implement sequence management with Sequence component and related composables
