+++ packages/core/src/composables/frame.js
@@ -1,0 +1,24 @@
+import { inject, computed } from 'vue'
+import { FRAME_KEY } from './keys.js'
+
+/**
+ * Get the current frame number.
+ * Must be used within a Pellicule render context.
+ *
+ * @returns {import('vue').ComputedRef<number>} Current frame number
+ *
+ * @example
+ * const frame = useFrame()
+ * const opacity = computed(() => frame.value / 30) // fade in over 1 second at 30fps
+ */
+export function useFrame() {
+  const frame = inject(FRAME_KEY)
+
+  if (frame === undefined) {
+    throw new Error(
+      'useFrame() must be used within a Pellicule render context'
+    )
+  }
+
+  return computed(() => frame.value)
+}
